move files to "duke" package
