+++ data/saveFile.txt
@@ -1,0 +1,5 @@
+[T][X] read book
+[D][ ] return book (by: Jun 06 2022)
+[E][ ] project meeting (at: zzz)
+[T][X] join sports club
+[T][ ] borrow book
