--- src/main/java/DukeException.java
@@ -1,8 +1,0 @@
-public class DukeException extends Exception {
-
-  public DukeException(String msg) {
-    super("\t-----------------------------------------------"
-        + "\n\t" + msg
-        + "\n\t-----------------------------------------------");
-  }
-}
