--- src/main/java/Task.java
@@ -1,31 +1,0 @@
-public class Task {
-  protected String description;
-  protected boolean isDone;
-
-  public Task(String description) {
-    this.description = description;
-    this.isDone = false;
-  }
-
-  public Task(String description, boolean isDone) {
-    this.description = description;
-    this.isDone = isDone;
-  }
-
-  public String getStatusIcon() {
-    return (isDone ? "X" : " "); // mark done task with X
-  }
-
-  public void markAsDone() {
-    this.isDone = true;
-  }
-
-  public void markAsUndone() {
-    this.isDone = false;
-  }
-
-  @Override
-  public String toString() {
-    return "[" + getStatusIcon() + "] " + description;
-  }
-}
