+++ src/main/java/duke/Duke.java
@@ -1,7 +1,4 @@
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Scanner;
+package duke;
 
 public class Duke {
     private TaskList taskList;
