+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,10 @@
+package duke;
+
+public class DukeException extends Exception {
+
+    public DukeException(String msg) {
+        super("\t-----------------------------------------------"
+            + "\n\t" + msg
+            + "\n\t-----------------------------------------------");
+    }
+}
